# openapi.yaml
openapi: 3.0.3
info:
  title: API Registry Service
  description: Centralized registry for API specifications and service discovery in the PDS 2.0 ecosystem
  version: 1.0.0
  contact:
    name: PDS 2.0 Team
servers:
  - url: https://api-registry.local
    description: Production server
  - url: http://localhost:3005
    description: Local development server
...
paths:
...
  /specs:
    get:
      summary: List all API specifications
      description: Retrieves a list of all registered API specifications
      operationId: listSpecs
      tags:
        - specs
      parameters:
        - name: service
          in: query
          description: Filter by service name
          required: false
          schema:
            type: string
        - name: version
          in: query
          description: Filter by API version
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of API specifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpecSummary'
    post:
      summary: Register a new API specification
      description: Registers a new OpenAPI specification for a service
      operationId: registerSpec
      tags:
        - specs
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecRegistration'
      responses:
        '201':
          description: API specification registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecDetail'
        '400':
          description: Invalid specification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Specification already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /services/{serviceId}:
    get:
      summary: Get service information
      description: Retrieves information about a specific service
      operationId: getService
      tags:
        - services
      parameters:
        - name: serviceId
          in: path
          description: The ID of the service
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetail'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "invalid_request"
        error_description:
          type: string
          example: "The request is missing a required parameter"
        error_uri:
          type: string
          example: "https://api-registry.local/docs/errors#invalid_request"
    SpecSummary:
      type: object
      properties:
        id:
          type: string
          example: "api-registry"
        title:
          type: string
          example: "API Registry Service"
        version:
          type: string
          example: "1.0.0"
        created_at:
          type: string
          format: date-time
          example: "2025-07-18T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-07-18T10:30:00Z"
    SpecDetail:
      type: object
      properties:
        id:
          type: string
          example: "api-registry"
        title:
          type: string
          example: "API Registry Service"
        version:
          type: string
          example: "1.0.0"
        description:
          type: string
          example: "Centralized registry for API specifications"
        spec:
          type: object
          description: "The complete OpenAPI specification"
        created_at:
          type: string
          format: date-time
          example: "2025-07-18T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-07-18T10:30:00Z"
    SpecRegistration:
      type: object
      properties:
        service_id:
          type: string
          example: "api-registry"
        spec:
          type: object
          description: "The complete OpenAPI specification"
      required:
        - service_id
        - spec
...
    ServiceDetail:
      type: object
      properties:
        id:
          type: string
          example: "api-registry"
        name:
          type: string
          example: "API Registry"
        description:
          type: string
          example: "Centralized registry for API specifications"
        url:
          type: string
          example: "https://api-registry.local"
        contact:
          type: object
          properties:
            name:
              type: string
              example: "PDS 2.0 Team"
            email:
              type: string
              example: "support@pds.local"
        versions:
          type: array
          items:
            type: object
            properties:
              version:
                type: string
                example: "1.0.0"
              status:
                type: string
                enum: [current, deprecated, beta]
                example: "current"
              deprecated_at:
                type: string
                format: date-time
                example: null
        status:
          type: string
          enum: [online, offline, maintenance]
          example: "online"
        metrics:
          type: object
          properties:
            uptime:
              type: number
              example: 99.99
            last_checked:
              type: string
              format: date-time
              example: "2025-07-18T12:30:00Z"